fix: split scoped factory save topology tests for lint
Move topology-classification describe block into a sibling test file so
useScopedFactoryDocumentSave.test.tsx stays under the noExcessiveLinesPerFile limit.
